gig_platform_backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
black==25.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Django==4.2.30
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gprof2dot==2025.4.14
idna==3.11
inflection==0.5.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
nltk==3.9.2
packaging==26.0
pathspec==1.0.4
pillow==11.3.0
platformdirs==4.4.0
psycopg2==2.9.12
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.5
tomli==2.4.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
